new/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pypdf
scikit-learn
numpy
mistralai
python-dotenv
# Using Mistral's embedding API instead of local models for better scalability
# and to reduce memory usage and deployment complexity